+++ tests/client/queued_processes_test.py
@@ -38,3 +38,20 @@ def test_queued_process(client):
     assert process.created_at_timestamp == 1592566922
     assert process.details['files'][0]['status'] == 'finished'
     assert process.details['files'][0]['name_original'] == 'en.json'
+
+
+@pytest.mark.vcr
+def test_queued_process_async_download(client):
+    """Tests fetching of a queued process for an async download
+    """
+    project_id = "6504960967ab53d45e0ed7.15877499"
+    process_id = "1efed4e6-461a-6d6e-a779-dea3ae8b7e9e"
+    process = client.queued_process(
+        project_id,
+        process_id
+    )
+
+    assert process.process_id == process_id
+    assert process.type == 'async-export'
+    assert process.status == 'finished'
+    assert 'https://lokalise-live-lok-s3-fss-export' in process.details['download_url']